Refactor: Standardize component folder naming and improve gitignore
  - Rename component folders to follow kebab-case convention:
    - PageHeader → page-header
    - Widget → widget
    - Loading → loading
    - BasePage.jsx → base-page/BasePage.jsx
  - Update all import paths across 17 files to reflect new structure
  - Fix EditIngredientModal tab switching state persistence issue
  - Improve FormField styling in Allergens modal for full width
  - Enhance gitignore with additional cache and temp file patterns
  - Maintain consistent folder structure for better code organization
