+++ frontend/src/components/widget/Widget.css
