+++ frontend/src/pages/events/Events.jsx
@@ -3,8 +3,8 @@ import React, { useEffect, useState, useMemo, useRef } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { FaCalendarAlt, FaUsers, FaExclamationTriangle, FaEuroSign } from 'react-icons/fa';
 import TableActions from '../../components/table/TableActions';
-import BasePage from '../../components/BasePage';
-import Widget from '../../components/Widget';
+import BasePage from '../../components/base-page/BasePage';
+import Widget from '../../components/widget';
 import Modal from '../../components/modal/Modal';
 import TabsModal from '../../components/tabs-modal/TabsModal';
 import { FormField, FormInput, FormTextarea, FormSelect } from '../../components/form/FormField';
