+++ frontend/src/pages/recipes/Recipes.jsx
@@ -5,8 +5,8 @@ import { StyleSheetManager } from 'styled-components';
 import isPropValid from '@emotion/is-prop-valid';
 import { FaPlus } from 'react-icons/fa';
 import TableActions from '../../components/table/TableActions';
-import BasePage from '../../components/BasePage';
-import Loading from '../../components/Loading';
+import BasePage from '../../components/base-page/BasePage';
+import Loading from '../../components/loading';
 import Modal from '../../components/modal/Modal';
 import { usePageState } from '../../hooks/usePageState';
 import { useSettings } from '../../context/SettingsContext';
